Replace local pub-sub with npm pub-sub for separation of concerns
