+++ app/scripts/Annotations2dTrack.js
@@ -1,10 +1,10 @@
 import * as PIXI from 'pixi.js';
+import createPubSub from 'pub-sub-es';
 
 import TiledPixiTrack from './TiledPixiTrack';
 
 // Services
 import { tileProxy } from './services';
-import createPubSub from './services/pub-sub';
 
 // Utils
 import { colorToHex, max, min } from './utils';
